# docs/api.yaml
openapi: 3.0.3
info:
  title: AI Diary API
  version: 1.0.0
  description: |
    AI 日记本后端接口定义。
    典型流程：
    1. 上传图片
    2. AI 分析图片
    3. 基于图片开启聊天
    4. 结束聊天并生成日记
    5. 查询历史日记
servers:
  - url: http://localhost:8000/api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateChatSessionRequest:
      type: object
      required: [image_id]
      properties:
        image_id:
          type: string
          format: uuid
        title:
          type: string
    CreateChatSessionResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        status:
          type: string
          example: active
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        stream:
          type: boolean
          default: false
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        sequence_no:
          type: integer
        created_at:
          type: string
          format: date-time
    EndChatResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        diary_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /chat/sessions:
    post:
      tags: [Chat]
      summary: 创建会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatSessionRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChatSessionResponse'
  /chat/sessions/{session_id}/messages:
    post:
      tags: [Chat]
      summary: 发送消息并获取 AI 回复
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: 非流式回复
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_message:
                    $ref: '#/components/schemas/ChatMessage'
                  assistant_message:
                    $ref: '#/components/schemas/ChatMessage'
...
  /chat/sessions/{session_id}/end:
    post:
      tags: [Chat]
      summary: 结束聊天并生成日记
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 已生成日记
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndChatResponse'
...